+++ propack-core/src/main/java/me/nelonn/propack/builder/task/FileProcessingException.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.nelonn.propack.core.builder.task;
+package me.nelonn.propack.builder.task;
 
 import org.jetbrains.annotations.NotNull;
