+++ propack-core/src/main/java/me/nelonn/propack/core/builder/task/GatherSourcesTask.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.nelonn.propack.core.builder.task.provided;
+package me.nelonn.propack.core.builder.task;
 
 import com.google.gson.JsonObject;
 import me.nelonn.flint.path.PathImpl;
@@ -26,8 +26,8 @@ import me.nelonn.propack.builder.file.JsonFile;
 import me.nelonn.propack.builder.file.RealFile;
 import me.nelonn.propack.builder.file.TextFile;
 import me.nelonn.propack.builder.task.TaskIO;
-import me.nelonn.propack.core.builder.task.AbstractTask;
-import me.nelonn.propack.core.builder.task.TaskBootstrap;
+import me.nelonn.propack.builder.task.AbstractTask;
+import me.nelonn.propack.builder.task.TaskBootstrap;
 import me.nelonn.propack.core.util.GsonHelper;
 import me.nelonn.propack.core.util.IOUtil;
 import me.nelonn.propack.core.util.LogManagerCompat;
